Fixed build error
